+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Coming soon` sections from README and documentation, as covered by backlogs.
   (https://github.com/gchq/coreax/pull/972)
+- **[BREAKING CHANGE]** Removed deprecated Coreset functionality - see v0.4.0's
+deprecations for more detail.
 
 ### Deprecated